pyspark/dataframe/sql-dataframe/requirements.txt
mysql-connector-python==8.0.31
numpy==1.21.6
pandas==1.3.5
protobuf==3.20.1
py4j==0.10.9.5
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
